RentRepository refactor, RentException added

+++ src/main/java/com/library/repository/rent/RentRepository.java
@@ -2,6 +2,8 @@ package com.library.repository.rent;
 
 import com.library.domain.book.Book;
 import com.library.domain.rent.Rent;
+import com.library.domain.user.User;
+import com.library.exception.RentException;
 import com.library.repository.book.BookDao;
 import com.library.repository.user.UserDao;
 import org.slf4j.Logger;
@@ -10,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Optional;
 
 @Component
@@ -24,33 +27,91 @@ public class RentRepository {
     @Autowired
     private BookDao bookDao;
 
-    public Rent rentBooks(Long userId, Long... bookId) {
+    public Rent rentBooks(Long userId, Long... bookId) throws RentException {
 
         Rent rent = new Rent(getReturnDate());
-        if (userDao.existsById(userId)) {
-            rent.setUser(userDao.findById(userId).get());
+        boolean doRent = false;
+
+        Optional<User> optionalUser = userDao.findById(userId);
+        if (optionalUser.isPresent()) {
+            User user = optionalUser.get();
+            rent.setUser(user);
 
             for (Long id : bookId) {
-                Optional<Book> book = bookDao.findById(id);
-                book.get().setRent(rent);
-                rent.getRentedBooks().add(book.get());
+                Optional<Book> optionalBook = bookDao.findById(id);
+                if (optionalBook.isPresent()) {
+                    Book book = optionalBook.get();
+                    book.setRent(rent);
+
+                    rent.getRentedBooks().add(book);
+
+                    doRent = true;
+
+                    LOGGER.info("Book {} rented", id);
+                } else {
+                    LOGGER.warn("No book with id {}", id);
+                }
             }
+        } else {
+            LOGGER.error("Failed to rent books");
+            throw new RentException(RentException.INVALID_USER_ID);
         }
-        return rentDao.save(rent);
+        if (doRent) {
+            LOGGER.info("Books rented successfully");
+            return rentDao.save(rent);
+        }
+        LOGGER.error("Failed to rent books");
+        throw new RentException(RentException.NOTHING_RENTED);
     }
 
-    public void returnRent(Long rentId) {
+    public void returnRent(Long rentId) throws RentException {
 
         Optional<Rent> rent = rentDao.findById(rentId);
-        rent.ifPresent(r -> r.getRentedBooks().forEach(book -> {
-            book.setRent(null);
-            book.setStatus(Book.Status.available);
-        }));
-        rentDao.deleteById(rentId);
+        if (rent.isPresent()) {
+            rent.get().getRentedBooks().forEach(book -> {
+                book.setRent(null);
+                book.setStatus(Book.Status.available);
+                LOGGER.info("Returned book {}", book.getId());
+            });
+            rentDao.deleteById(rentId);
+            LOGGER.info("Returned rent {}", rentId);
+        } else {
+            LOGGER.error("Failed to return rent");
+            throw new RentException(RentException.INVALID_RENT_ID);
+        }
+    }
 
+    public void returnBook(Long... bookId) throws RentException {
+
+        for (Long id : bookId) {
+            Optional<Book> optional = bookDao.findById(id);
+            if (optional.isPresent()) {
+                Book book = optional.get();
+                book.setStatus(Book.Status.available);
+
+                Rent rent = book.getRent();
+                List<Book> rentedBooks = rent.getRentedBooks();
+                rentedBooks.remove(book);
+
+                book.setRent(null);
+
+                deleteEmptyRent(rent);
+
+                LOGGER.info("Returned book {}", id);
+            } else {
+                LOGGER.error("Failed to return book");
+                throw new RentException(RentException.INVALID_BOOK_ID);
+            }
+        }
     }
 
     private LocalDate getReturnDate() {
         return LocalDate.now().plusMonths(2);
+    }
+
+    private void deleteEmptyRent(Rent rent) {
+        if (rent.getRentedBooks().isEmpty()) {
+            rentDao.delete(rent);
+        }
     }
 }
